change all considerable uint32 to uint64

+++ structures.h
@@ -90,7 +90,7 @@ struct s_dir_entry2 {
 	uint8	name_len;		/* Name length */
 	uint8	file_type;
 	char	name[EXT2_NAME_LEN];	/* File name */
-	uint32 block_number;
+	uint64 block_number;
 	uint32 offset;
 };
